use ints

+++ constexpr_bench.cpp
@@ -1,6 +1,6 @@
 #include "bench.h"
 
-constexpr int64_t get_bucket_constexpr(const int64_t value, size_t index = NUM_BUCKETS - 1) {
+constexpr int get_bucket_constexpr(const int64_t value, size_t index = NUM_BUCKETS - 1) {
 
     //return ~((value - thresholds[index]) >> 63) + get_bucket_constexpr(value, index - 1);
     return index == 0
