+++ template_bench.cpp
@@ -3,7 +3,7 @@
 // Utility to generate masks and index calculation at compile time
 template <size_t N>
 struct BucketIndexHelper {
-    static int64_t get_bucket_index(const int64_t value) {
+    static int get_bucket_index(const int64_t value) {
         int64_t mask = ~((value - thresholds[N-1]) >> 63);
         return (mask & 1) + BucketIndexHelper<N-1>::get_bucket_index(value);
     }
@@ -19,13 +19,13 @@ struct BucketIndexHelper<1> {
 
 template <>
 struct BucketIndexHelper<0> {
-    static int64_t get_bucket_index(int64_t) {
+    static int get_bucket_index(int64_t) {
         return 0;
     }
 };
 
 template <size_t N>
-int64_t get_bucket_index(const int64_t value) {
+int get_bucket_index(const int64_t value) {
     //static_assert(N <= thresholds.size(), "N must not exceed the size of thresholds");
     return BucketIndexHelper<N>::get_bucket_index(value);
 }
